refactor: streamline codebase to single-file architecture
- Consolidated distributed codebase into clean single-file main.rs
- Moved legacy files to legacy/ directory for reference
- Added Makefile for build automation and cross-compilation
- Simplified dependencies and removed unnecessary complexity
- Maintained all core functionality: CPU, memory, disk, network monitoring
- Preserved web dashboard with responsive design improvements
- Updated README.md with simplified architecture documentation

This creates a cleaner, more maintainable codebase focused on core functionality.
